fix(board): resolve_showdown の invested_sum を u64 化 (R32 Bug A2)
total_invested は u32 だが 6 人× 800_000_000 など累積で u32::MAX を超えうる。
sum() を u64 経由で計算し pot_total との比較も u64 同士に揃えることで
warn の誤発火と整数 overflow を防ぐ。

+++ src-tauri/src/domain/board.rs
@@ -383,8 +383,8 @@ impl TexasHoldemBoard {
             return;
         }
 
-        let invested_sum: u32 = self.players.iter().map(|p| p.total_invested).sum();
-        let pot_total = self.total_pot();
+        let invested_sum: u64 = self.players.iter().map(|p| p.total_invested as u64).sum();
+        let pot_total = self.total_pot() as u64;
         if invested_sum != pot_total {
             tracing::warn!(
                 "resolve_showdown: total_invested sum ({}) != total_pot ({}); \
